Simplify "open or create" retrieval of default value; fix regressions
I recently refactored the denote-title-prompt to use 'completing-read'
and missed the part of setting the 'denote-title-prompt-current-default'.
This is now restored.

I am also simplifying how we are working with the minibuffer histories
to (i) not pollute the denote--file-history with unrelated strings
and (ii) to not have to deal with 'let' bound values in places where I
would rather not have them.

I delved into this in response to the message sent by Aleksandr
Vityazev on the mailing list:
<https://lists.sr.ht/~protesilaos/denote/%3C87a5qevq8i.fsf%40gmail.com%3E>.