Make denote-update-dired-buffers also revert current Dired
This means that we revert the current Dired buffer even if it is not
part of the 'denote-directory'. We do so because the "rename" commands
we provide can be used everywhere on the filesystem.

Thanks to babusri for bringing this matter to my attention. This was
done in issue 203 on the GitHub mirror:
<https://github.com/protesilaos/denote/issues/203>.