Fix for denote-rename-file when signature has ='s for spaces.
When a signature is created and spaces are used, the spaces are
converted into `='s in the function =denote--slug-put-equals=.

When a denote file with a signature is renamed, the signature
is presented to the user during the renaming process with all
of the `='s intact.  If the user then accepts the signaturre
as-is or changes other parts of the signature but leaves the
`='s in place, the `='s are then removed by the function
denote--slug-no-punct=, which uses the constant regular
expression =denote-excluded-punctuation-regexp=, which
includes the `='.  The resulting new filename is thus missing
all `='s that were in the original, although the user intended
them to stay (by not removing them).

This fix simply replaces all `='s with ` 's (a single space) prior
to presenting the signature to the user during the renaming process.
The user can then accept the signature as-is, or change any part of
it, and the spaces will subsequently be converted into `='s by
denote--slug-put-equals= and will not be lost.