Make denote--edit-front-matter-p check if front matter is non-nil
We do not want to try to rewrite front matter if the given file type has
a nil value for its front matter. Furthermore, we do not want to make
the inclusion of front matter mandatory. Front matter is not a
requirement for Denote, as the file name is our point of reference.

Thanks to Glenna D. for reporting an error produced by
'denote-dired-rename-marked-files-with-keywords' in a scenario where
the front matter would be nil. This was done via a private channel
and the information is shared with permission.