Fix edge-cases when inserting missing links via org-dblock
To reproduce the problem: open a new note and while it is empty, try
to insert and populate an org dblock of type "denote-missing-links".

This leads to the following stack-trace:

```
Debugger entered--Lisp error: (wrong-type-argument stringp 65)
  file-name-nondirectory(65)
  file-name-extension(65 :period)
  denote-get-file-extension(65)
  denote-get-file-extension-sans-encryption(65)
  denote-filetype-heuristics(65)
  denote-link-description-with-signature-and-title(65 nil)
  denote--link-get-description(65)
  #f(compiled-function (file) #<bytecode -0x3aa12408d80ef11>)(65)
  mapc(#f(compiled-function (file) #<bytecode -0x3aa12408d80ef11>) "All links matching ‘reference/2024012’ are present")
  denote-link--prepare-links("All links matching ‘reference/2024012’ are present" org nil :no-other-sorting)
  denote-link--insert-links("All links matching ‘reference/2024012’ are present" org nil :no-other-sorting)
  org-dblock-write:denote-missing-links((:name "denote-missing-links" :regexp "reference/2024012" :indentation-column 0 :content #("\n" 0 1 (org-fold-core-fontified t fontified t))))
```

There are two issues in the underlying code:

1. When `denote-org-extras-dblock--get-missing-links` finds 0
missing-links, it returns with a string (return value of `message`)
instead of with an empty list.

2. When there are 0 pre-existing denote links in the current
buffer, `(denote-link--expand-identifiers
denote-org-link-in-context-regexp)` returns nil. This sets
`linked-files` value to `nil` and short-circuits the `if-let` when it
should not.

This commit fixes these two issues.