Reinstate relative file paths in 'denote-file-prompt'
This affects every Denote command that prompts for a file (e.g. to
link to it).

Previous versions of Denote had this feature, but it was depending on
a project.el code that did not do exactly what we were expecting.

With Jean-Philippe Gagné Guay we tried to relativise file paths using
display properties, but could not get the completion highlights to
work correctly. This was discussed in pull request 239, with input
from Mark Olson: <https://github.com/protesilaos/denote/pull/239>. A
similar request for relative file names was considered in issue 233
with the participation of Vick (VicZz):
<https://github.com/protesilaos/denote/issues/233>.