Make denote-rename-file and denote-dired-rename-files prompt for signature, if it exists
This is a case where we do not want to unconditionally remove the
signature when 'denote-prompts' does not include a signature prompt.
Signatures only exist in the file name, not any front matter, so we
want to give the user the chance to decide what to do with an existing
one while renaming the file.

Files without a signature are not affected by this change.