Make denote-file-prompt more robust
Because we do not force a match, we need to check if the user's input
matches a file uniquely. Otherwise we return the input as-is, which is
then used by 'denote-open-or-create' (and related).

To make sure that the latest input is used and to not mess up with the
minibuffer history, we now have a new internal variable that handles
this case.