Revise the denote-file-prompt to improve the user experience
The goal is to decouple file selection from the logic of the "... or
create" commands we have. We want to require a match to a file by
default and to only have a fallback "or create" action in those cases
where it is requested.

The changes I make here apply to the "... or create" commands, though
they also affect 'denote-link' and related. When we prompt to link to
a file, we require a match, such that the user cannot exit the
minibuffer with a string that does not expand to an existing file.
This was an error we were not handling. It was also an easy mistake to
make.

This follows the discussion with Jean-Philippe Gagné Guay and Alan
Schmitt in issue 339: <https://github.com/protesilaos/denote/issues/339>.