Make denote-title-prompt prepopulate the minibuffer if denote-title-prompt-current-default
This makes it easier to edit the text after calling 'denote-open-or-create' (and related).
We do not want to skip the title prompt completely, because the user
may not necessarily search for a file using a query that is a
perfectly valid title. For example, with the 'orderless' package set
up, they may type "title test" to match "This is a test title".