Change the dummy regexp for front matter retrieval to something more self-evident
This was originally contributed by Jean-Philippe Gagné Guay in pull
request 480: <https://github.com/protesilaos/denote/pull/480>. The
idea is to use this regular expression which will 99% not exist in a
file to handle the case where the user does not want to get the value
of the given entry. The alternative would be to rewrite more parts of
the code, so this feels safer and is easier.