Make denote-journal-extras--filename-date-regexp more robust
We want it to work when the journal files are inside a directory and
we also need it to account for variations of denote-file-name-components-order.

Thanks to Samuel Flint for informing me about the previous
implementation not working inside of a journal subdirectory. This was
done in issue 485: <https://github.com/protesilaos/denote/issues/485>.