DEPRECATE denote-link-with-signature
The denote-link command already inserts a signature if it is there,
plus we have the denote-link-description-format, which can be set to
whatever style the user prefers. This command I am deprecating was
basically doing what a user gets as soon as they type = in the
minibuffer (which will match files that have a signature).