Add an obsolete alias for denote-slug-hyphenate
In principle, we should not be doing this for private symbols, but it
helps with transitioning old code to the new normal. We have had this
practice for a long time now (and I eventually remove those warnings a
few releases afterwards).

Thanks to bdillahu for reporting the issue with some sample code in
the manual and to Jean-Philippe Gagné Guay for confirming the problem.
This was done in issue 538: <https://github.com/protesilaos/denote/issues/538>.