Check for empty sequence in 'denote-sequence-get-all-files-with-prefix'
This prevents an error when calling 'denote-sequence-split' with an
empty string.  In particular, this happens when
'denote-sequence-get-relative' is called with TYPE 'siblings' on a
top-level sequence.  'denote-sequence-get-all-sequences-with-prefix' is
apparently not used in a context where it would receive an empty
sequence, but check for empty sequences there as well for consistency.