Make denote-grep-query-prompt return a string, not assuming it is part of 'interactive'
This is how we do it with all our prompts. In practice, it is used in
an interactive spec though it might be used outside of one (also
considering how users may choose to use our functions).

This builds on top of commit f70c96c, which was contributed by Lucas
Quintana in pull request 571: <https://github.com/protesilaos/denote/pull/571>.

Lucas has assigned copyright to the Free Software Foundation.