Make denote-sort-dired use a unique buffer each time
When we were passing the directory as (dired (cons DIRECTORY FILES))
we were inadvertently setting up Dired to re-use the buffer it
produced. Whereas we always want to have a new one for distinct
searches.

This is how we were doing it in the past but we had issues with the
revert-buffer function. Maybe we are good now.