Make the denote-directory function obsolete
I have gone back and forth on this issue. Jean-Philippe Gagné Guay
convinced me that it is better to deprecate it because this way the
users will be more informed about what is happening.

This is done in a comment and was then reminded it of it again in pull
request 632:

- <https://github.com/protesilaos/denote/commit/947f0be070c686f2b253498d6eb65ac2abcaee8e>.
- <https://github.com/protesilaos/denote/pull/632>